ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
APScheduler==3.10.4
prometheus-client==0.19.0
